I'm still learning!
I have been doing this the hard way by taking Nexus .savs and using GitHubs B3HM (@c0dycode) and profile editors (@Apocalyptech) and trying to combine things cohesively for myself and others. Long term goals are to learn well enough to do this and develop an editor that will be user friendly and able to run on chromebooks and a Xbox One.